--- core/src/main/java/com/github/can019/Main.java
@@ -1,7 +1,0 @@
-package com.github.can019;
-
-public class Main {
-    public static void main(String[] args) {
-        System.out.println("Hello world!");
-    }
-}
