+++ core/src/main/java/com/github/can019/base/core/CoreApplication.java
@@ -1,0 +1,10 @@
+package com.github.can019.base.core;
+
+import org.springframework.boot.autoconfigure.SpringBootApplication;
+
+@SpringBootApplication
+public class CoreApplication {
+    public static void main(String[] args) {
+        System.out.println("Hello world!");
+    }
+}
