+++ core/src/main/java/com/github/can019/base/core/entity/BaseTimeEntity.java
@@ -1,0 +1,20 @@
+package com.github.can019.base.core.entity;
+
+import jakarta.persistence.MappedSuperclass;
+import lombok.Getter;
+import org.hibernate.annotations.CreationTimestamp;
+import org.hibernate.annotations.UpdateTimestamp;
+
+import java.time.LocalDateTime;
+
+@Getter
+@MappedSuperclass
+public abstract class BaseTimeEntity {
+    @CreationTimestamp
+    private LocalDateTime createdTime;
+
+    @UpdateTimestamp
+    private LocalDateTime updatedTime;
+
+
+}
